+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Allow the notifications details panel to be moved to the bottom of the screen
 - Add a new way to turn notifications into a task with default parameters
 - Add Google Drive integration
+- Prevent user registration from blck listed domains
 
 ### Changed